separated files for DragDrapBoard component and the page using the component.
