+++ src/pages/dragdrop/types/index.ts
@@ -1,0 +1,10 @@
+import { UniqueIdentifier } from '@dnd-kit/core';
+
+export interface DDType {
+  id: UniqueIdentifier
+}
+
+export interface DDCategory<T>{
+  value: T;
+  label: string;
+}
